requirements.txt
dj-database-url==0.5.0
Django==3.1.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.2.0
django-crispy-forms==1.9.2
django-djconfig==0.10.0
django-gm2m==1.1.1
django-haystack==3.0
django-heroku==0.3.1
django-infinite-scroll-pagination==1.1.0
django-spirit==0.11.0
gbulb==0.6.1
gitdb==4.0.5
GitPython==3.1.9
gunicorn==20.0.4
Pillow==7.2.0
poyo==0.5.0
psycopg2==2.8.5
pycairo==1.20.0
pycodestyle==2.6.0
pycrypto==2.6.1
pygobject==3.26.1
python-apt===1.6.5-ubuntu0.3-elementary16-ubuntu5.1.5.1
python-dateutil==2.8.1
python-decouple==3.3
python-slugify==4.0.1
pytz==2020.1
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.24.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
six==1.15.0
smmap==3.0.4
soupsieve==2.0.1
sqlparse==0.4.1
text-unidecode==1.3
toga-core==0.3.0.dev24
toga-gtk==0.3.0.dev24
toml==0.10.1
travertino==0.1.3
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.25.11
whitenoise==5.2.0
